requirements_production.txt
# Django e dependências principais
Django==5.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.0

# Segurança e variáveis de ambiente
python-dotenv==1.1.1

# Banco de dados MySQL para PythonAnywhere
PyMySQL==1.1.1

# APIs de IA
openai==1.54.4
anthropic==0.34.2

# Utilitários
requests==2.31.0

# Para produção no PythonAnywhere
whitenoise==6.6.0
